HS-1484: Consolidate tag creation at Inventory service (#1226)
* HS-1484: Consolidate tag creation at Inventory service

* HS-1484: fix tags updating to inventory service

* HS-1484: Remove unused code

* HS-1484: Progagate tag updates to Alert service

* HS-1484: Add monitoring policy ids as jsonb in tags table

* HS-1484: Fix test failures

* HS-1484: Fix tags for alerts, add tests

also handle review comments

* HS-1484: Persist tags independent of monitor policies

* HS-1484: Remove test

* HS-1484: Remove progapation of tags from inventory to alerts

* HS-1484: Update existing tags

Remove tags if no monitoring policy exist
Prevent duplicates in policy ids

* HS-1484: Handle review comments

* HS-1484: handle upgrades on tag table

+++ alert/src/main/java/org/opennms/horizon/alertservice/service/TagService.java
@@ -28,17 +28,18 @@
 
 package org.opennms.horizon.alertservice.service;
 
-import java.util.List;
-
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.opennms.horizon.alert.tag.proto.TagProto;
 import org.opennms.horizon.alertservice.db.entity.Tag;
+import org.opennms.horizon.alertservice.db.repository.MonitorPolicyRepository;
 import org.opennms.horizon.alertservice.db.repository.TagRepository;
 import org.opennms.horizon.alertservice.mapper.TagMapper;
 import org.opennms.horizon.shared.common.tag.proto.TagOperationList;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import java.util.List;
 
 @RequiredArgsConstructor
 @Slf4j
@@ -46,7 +47,9 @@ import lombok.extern.slf4j.Slf4j;
 public class TagService {
     private final TagRepository tagRepository;
     private final TagMapper tagMapper;
+    private final MonitorPolicyRepository monitorPolicyRepository;
 
+    @Transactional
     public void insertOrUpdateTags(TagOperationList list) {
         list.getTagsList().forEach( tagOp -> {
             switch (tagOp.getOperation()) {
@@ -72,7 +75,7 @@ public class TagService {
                     .ifPresent(tag -> {
                         int oldSize = tag.getNodeIds().size();
                         tagOp.getNodeIdList().forEach(id -> tag.getNodeIds().remove(id));
-                        if(tag.getNodeIds().isEmpty()) {
+                        if(tag.getNodeIds().isEmpty() && tag.getPolicies().isEmpty()) {
                             tagRepository.deleteById(tag.getId());
                             log.info("deleted tag {}", tagOp);
                         } else {
@@ -86,6 +89,6 @@ public class TagService {
 
     public List<TagProto> listAllTags(String tenantId) {
         return tagRepository.findByTenantId(tenantId)
-            .stream().map(tag -> tagMapper.map(tag)).toList();
+            .stream().map(tagMapper::map).toList();
     }
 }
